broke up quadrotor Vector classes into their own files
